refactor: reorganize ShortenedLinkPage as parent component to CopyLink and InputField components
